zynqmp_pmufw: Fix MISRA C Rule 10.3 violation
Resolved implicit conversion warning by explicitly
casting the enum value MEM_RANGE_WRITE_ACCESS to
u32 in variable assignment.

Acked-by: Jay Buddhabhatti <jay.buddhabhatti@amd.com>

+++ lib/sw_apps/zynqmp_pmufw/src/pm_core.c
@@ -1052,7 +1052,7 @@ static void PmFpgaRead(const PmMaster *const master,
 	UINTPTR Address = ((u64)AddrHigh << 32U)|AddrLow;
 #if defined(ENABLE_MEM_RANGE) && defined(ENABLE_MEM_RANGE_PM_FPGA_READ_BACK)
 	u32 length = Reg_Numframes * 4U;
-	u32 access = MEM_RANGE_WRITE_ACCESS;
+	u32 access = (u32)MEM_RANGE_WRITE_ACCESS;
 #endif
 #if defined(ENABLE_FPGA_READ_CONFIG_REG)
    u32 RspBuf __attribute__ ((aligned(64)));
